requirements/local.txt
asgiref==3.6.0
Django==4.1.6
django-ckeditor==6.5.1
django-js-asset==2.0.0
django-model-utils==4.3.1
Pillow==9.4.0
psycopg2-binary==2.9.5
sqlparse==0.4.3
Unipath==1.1